Move skipped tests
