+++ src/test/java/org/opentripplanner/updater/trip/TimetableSnapshotSourceTest.java
@@ -2,7 +2,6 @@ package org.opentripplanner.updater.trip;
 
 import static com.google.transit.realtime.GtfsRealtime.TripDescriptor.ScheduleRelationship.ADDED;
 import static com.google.transit.realtime.GtfsRealtime.TripDescriptor.ScheduleRelationship.CANCELED;
-import static com.google.transit.realtime.GtfsRealtime.TripDescriptor.ScheduleRelationship.SCHEDULED;
 import static com.google.transit.realtime.GtfsRealtime.TripUpdate.StopTimeUpdate.ScheduleRelationship.SKIPPED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -29,7 +28,6 @@ import org.junit.jupiter.api.Test;
 import org.opentripplanner.ConstantsForTests;
 import org.opentripplanner.TestOtpModel;
 import org.opentripplanner._support.time.ZoneIds;
-import org.opentripplanner.framework.i18n.NonLocalizedString;
 import org.opentripplanner.framework.time.ServiceDateUtils;
 import org.opentripplanner.model.PickDrop;
 import org.opentripplanner.model.Timetable;
@@ -307,219 +305,6 @@ public class TimetableSnapshotSourceTest {
         newTimetableScheduled.getTripIndex(modifiedTripId),
         "New trip should not be found in scheduled time table"
       );
-    }
-  }
-
-  @Nested
-  class Scheduled {
-
-    @Test
-    public void scheduled() {
-      // GIVEN
-
-      String scheduledTripId = "1.1";
-
-      var builder = new TripUpdateBuilder(
-        scheduledTripId,
-        SERVICE_DATE,
-        SCHEDULED,
-        transitModel.getTimeZone()
-      )
-        .addDelayedStopTime(1, 0)
-        .addDelayedStopTime(2, 60, 80)
-        .addDelayedStopTime(3, 90, 90);
-
-      var tripUpdate = builder.build();
-
-      var updater = defaultUpdater();
-
-      // WHEN
-      updater.applyTripUpdates(
-        TRIP_MATCHER_NOOP,
-        REQUIRED_NO_DATA,
-        DIFFERENTIAL,
-        List.of(tripUpdate),
-        feedId
-      );
-
-      // THEN
-      final TimetableSnapshot snapshot = updater.getTimetableSnapshot();
-
-      final FeedScopedId tripId = new FeedScopedId(feedId, scheduledTripId);
-      final Trip trip = transitModel.getTransitModelIndex().getTripForId().get(tripId);
-      final TripPattern originalTripPattern = transitModel
-        .getTransitModelIndex()
-        .getPatternForTrip()
-        .get(trip);
-
-      final Timetable originalTimetableForToday = snapshot.resolve(
-        originalTripPattern,
-        SERVICE_DATE
-      );
-      final Timetable originalTimetableScheduled = snapshot.resolve(originalTripPattern, null);
-
-      assertNotSame(originalTimetableForToday, originalTimetableScheduled);
-
-      final int originalTripIndexScheduled = originalTimetableScheduled.getTripIndex(tripId);
-      assertTrue(
-        originalTripIndexScheduled > -1,
-        "Original trip should be found in scheduled time table"
-      );
-      final TripTimes originalTripTimesScheduled = originalTimetableScheduled.getTripTimes(
-        originalTripIndexScheduled
-      );
-      assertFalse(
-        originalTripTimesScheduled.isCanceledOrDeleted(),
-        "Original trip times should not be canceled in scheduled time table"
-      );
-      assertEquals(RealTimeState.SCHEDULED, originalTripTimesScheduled.getRealTimeState());
-
-      final int originalTripIndexForToday = originalTimetableForToday.getTripIndex(tripId);
-      assertTrue(
-        originalTripIndexForToday > -1,
-        "Original trip should be found in time table for service date"
-      );
-      final TripTimes originalTripTimesForToday = originalTimetableForToday.getTripTimes(
-        originalTripIndexForToday
-      );
-      assertEquals(RealTimeState.UPDATED, originalTripTimesForToday.getRealTimeState());
-      assertEquals(0, originalTripTimesForToday.getArrivalDelay(0));
-      assertEquals(0, originalTripTimesForToday.getDepartureDelay(0));
-      assertEquals(60, originalTripTimesForToday.getArrivalDelay(1));
-      assertEquals(80, originalTripTimesForToday.getDepartureDelay(1));
-      assertEquals(90, originalTripTimesForToday.getArrivalDelay(2));
-      assertEquals(90, originalTripTimesForToday.getDepartureDelay(2));
-    }
-
-    @Test
-    public void scheduledTripWithSkippedAndNoData() {
-      // GIVEN
-
-      String scheduledTripId = "1.1";
-
-      var builder = new TripUpdateBuilder(
-        scheduledTripId,
-        SERVICE_DATE,
-        SCHEDULED,
-        transitModel.getTimeZone()
-      )
-        .addNoDataStop(1)
-        .addSkippedStop(2)
-        .addNoDataStop(3);
-
-      var tripUpdate = builder.build();
-
-      var updater = defaultUpdater();
-
-      // WHEN
-      updater.applyTripUpdates(
-        TRIP_MATCHER_NOOP,
-        REQUIRED_NO_DATA,
-        DIFFERENTIAL,
-        List.of(tripUpdate),
-        feedId
-      );
-
-      // THEN
-      final TimetableSnapshot snapshot = updater.getTimetableSnapshot();
-
-      // Original trip pattern
-      {
-        final FeedScopedId tripId = new FeedScopedId(feedId, scheduledTripId);
-        final Trip trip = transitModel.getTransitModelIndex().getTripForId().get(tripId);
-        final TripPattern originalTripPattern = transitModel
-          .getTransitModelIndex()
-          .getPatternForTrip()
-          .get(trip);
-
-        final Timetable originalTimetableForToday = snapshot.resolve(
-          originalTripPattern,
-          SERVICE_DATE
-        );
-        final Timetable originalTimetableScheduled = snapshot.resolve(originalTripPattern, null);
-
-        assertNotSame(originalTimetableForToday, originalTimetableScheduled);
-
-        final int originalTripIndexScheduled = originalTimetableScheduled.getTripIndex(tripId);
-        assertTrue(
-          originalTripIndexScheduled > -1,
-          "Original trip should be found in scheduled time table"
-        );
-        final TripTimes originalTripTimesScheduled = originalTimetableScheduled.getTripTimes(
-          originalTripIndexScheduled
-        );
-        assertFalse(
-          originalTripTimesScheduled.isCanceledOrDeleted(),
-          "Original trip times should not be canceled in scheduled time table"
-        );
-        assertEquals(RealTimeState.SCHEDULED, originalTripTimesScheduled.getRealTimeState());
-
-        final int originalTripIndexForToday = originalTimetableForToday.getTripIndex(tripId);
-        assertTrue(
-          originalTripIndexForToday > -1,
-          "Original trip should be found in time table for service date"
-        );
-        final TripTimes originalTripTimesForToday = originalTimetableForToday.getTripTimes(
-          originalTripIndexForToday
-        );
-        assertTrue(
-          originalTripTimesForToday.isDeleted(),
-          "Original trip times should be deleted in time table for service date"
-        );
-        // original trip should be deleted
-        assertEquals(RealTimeState.DELETED, originalTripTimesForToday.getRealTimeState());
-      }
-
-      // New trip pattern
-      {
-        final TripPattern newTripPattern = snapshot.getRealtimeAddedTripPattern(
-          new FeedScopedId(feedId, scheduledTripId),
-          SERVICE_DATE
-        );
-        assertNotNull(newTripPattern, "New trip pattern should be found");
-
-        final Timetable newTimetableForToday = snapshot.resolve(newTripPattern, SERVICE_DATE);
-        final Timetable newTimetableScheduled = snapshot.resolve(newTripPattern, null);
-
-        assertNotSame(newTimetableForToday, newTimetableScheduled);
-
-        assertTrue(newTripPattern.canBoard(0));
-        assertFalse(newTripPattern.canBoard(1));
-        assertTrue(newTripPattern.canBoard(2));
-
-        assertEquals(new NonLocalizedString("foo"), newTripPattern.getTripHeadsign());
-        assertEquals(
-          newTripPattern.getOriginalTripPattern().getTripHeadsign(),
-          newTripPattern.getTripHeadsign()
-        );
-
-        final int newTimetableForTodayModifiedTripIndex = newTimetableForToday.getTripIndex(
-          scheduledTripId
-        );
-        assertTrue(
-          newTimetableForTodayModifiedTripIndex > -1,
-          "New trip should be found in time table for service date"
-        );
-
-        var newTripTimes = newTimetableForToday.getTripTimes(newTimetableForTodayModifiedTripIndex);
-        assertEquals(RealTimeState.UPDATED, newTripTimes.getRealTimeState());
-
-        assertEquals(
-          -1,
-          newTimetableScheduled.getTripIndex(scheduledTripId),
-          "New trip should not be found in scheduled time table"
-        );
-
-        assertEquals(0, newTripTimes.getArrivalDelay(0));
-        assertEquals(0, newTripTimes.getDepartureDelay(0));
-        assertEquals(0, newTripTimes.getArrivalDelay(1));
-        assertEquals(0, newTripTimes.getDepartureDelay(1));
-        assertEquals(0, newTripTimes.getArrivalDelay(2));
-        assertEquals(0, newTripTimes.getDepartureDelay(2));
-        assertTrue(newTripTimes.isNoDataStop(0));
-        assertTrue(newTripTimes.isCancelledStop(1));
-        assertTrue(newTripTimes.isNoDataStop(2));
-      }
     }
   }
